+++ flutter_local_notifications/macos/flutter_local_notifications/Sources/flutter_local_notifications/FlutterLocalNotificationsPlugin.swift
@@ -182,6 +182,8 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
             requestPermissions(call, result)
         case "checkPermissions":
             checkPermissions(call, result)
+        case "openAppNotificationSettings":
+            openAppNotificationSettings(result)
         case "getNotificationAppLaunchDetails":
             getNotificationAppLaunchDetails(result)
         case "cancel":
@@ -598,6 +600,32 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
 
             result(dict)
         }
+    }
+
+    func openAppNotificationSettings(_ result: @escaping FlutterResult) {
+        let notificationsPath = "x-apple.systempreferences:com.apple.Notifications-Settings.extension"
+        
+        guard let bundleId = Bundle.main.bundleIdentifier else { return }
+        var openedAppNotificationSettings = false
+        if let url = URL(string: "\(notificationsPath)?id=\(bundleId)") {
+            openedAppNotificationSettings = NSWorkspace.shared.open(url)
+        }
+
+        if(openedAppNotificationSettings) {
+            result(openedAppNotificationSettings)
+            return;
+        }
+
+        openNotificationSettings(result)
+    }
+
+    func openNotificationSettings(_ result: @escaping FlutterResult) {
+        guard let url = URL(string: "x-apple.systempreferences:com.apple.preference.notifications") else {
+            result(false)
+            return
+        }
+
+        result(NSWorkspace.shared.open(url))
     }
 
     func getIdentifier(fromArguments arguments: [String: AnyObject]) -> String {