+++ flutter_local_notifications_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [12.2.0]
+
+* Added `openAppNotificationSettings()` method. When implemented, this opens the system settings UI where the user can manage notification permissions for the app. Thanks to the PR from [hamadcc](https://github.com/hamadcc)
+
 ## [12.1.0]
 
 * Added `NotificationResponseType.notificationDismissed` enum value to support handling when a notification has been dismissed. Thanks to the PR from [Vasily Laushkin](https://github.com/vlaushkin)